+++ Cargo.toml
@@ -105,6 +105,7 @@ members = [
     "sdk/macro",
     "sdk/package-metadata-macro",
     "sdk/program",
+    "sdk/program-memory",
     "send-transaction-service",
     "short-vec",
     "stake-accounts",
@@ -383,6 +384,7 @@ solana-perf = { path = "perf", version = "=2.1.0" }
 solana-poh = { path = "poh", version = "=2.1.0" }
 solana-poseidon = { path = "poseidon", version = "=2.1.0" }
 solana-program = { path = "sdk/program", version = "=2.1.0", default-features = false }
+solana-program-memory = { path = "sdk/program-memory", version = "=2.1.0" }
 solana-program-runtime = { path = "program-runtime", version = "=2.1.0" }
 solana-program-test = { path = "program-test", version = "=2.1.0" }
 solana-pubsub-client = { path = "pubsub-client", version = "=2.1.0" }
