+++ programs/bpf_loader/Cargo.toml
@@ -21,6 +21,7 @@ solana-curve25519 = { workspace = true }
 solana-log-collector = { workspace = true }
 solana-measure = { workspace = true }
 solana-poseidon = { workspace = true }
+solana-program-memory = { workspace = true }
 solana-program-runtime = { workspace = true }
 solana-sdk = { workspace = true }
 solana-timings = { workspace = true }
