+++ sdk/program-memory/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "solana-program-memory"
+description = "Basic low-level memory operations for Solana."
+documentation = "https://docs.rs/solana-program-memory"
+version = { workspace = true }
+authors = { workspace = true }
+repository = { workspace = true }
+homepage = { workspace = true }
+license = { workspace = true }
+edition = { workspace = true }
+
+[dependencies]
+num-traits.workspace = true
+
+[package.metadata.docs.rs]
+targets = ["x86_64-unknown-linux-gnu"]
+
+[target.'cfg(target_os = "solana")'.dependencies]
+solana-define-syscall = { workspace = true }
