+++ sdk/program/src/account_info.rs
@@ -3,8 +3,9 @@
 use {
     crate::{
         clock::Epoch, debug_account_data::*, entrypoint::MAX_PERMITTED_DATA_INCREASE,
-        program_error::ProgramError, program_memory::sol_memset, pubkey::Pubkey,
+        program_error::ProgramError, pubkey::Pubkey,
     },
+    solana_program_memory::sol_memset,
     std::{
         cell::{Ref, RefCell, RefMut},
         fmt,
